Show preview icon for "Add Map Marker" action

+++ OsmAnd/src/net/osmand/plus/base/containers/ShiftedBitmap.kt
@@ -2,4 +2,4 @@ package net.osmand.plus.base.containers
 
 import android.graphics.Bitmap
 
-data class ShiftedBitmap(val bitmap: Bitmap, val marginX: Int, val marginY: Int)
+data class ShiftedBitmap(val bitmap: Bitmap, val marginX: Float, val marginY: Float)
